refactor(tests): split data import to utils file
