+++ package.json
@@ -59,9 +59,9 @@
   },
   "ava": {
     "files": [
-      "tests/**/*.js"
+      "tests/**/*.spec.js"
     ],
-    "concurrency": 5,
+    "concurrency": 1,
     "failFast": true,
     "failWithoutAssertions": false,
     "verbose": true,
