+++ tests/init.spec.js
@@ -15,36 +15,12 @@ const dashboard = require('../src/models/dashboard');
 const reset = require('../src/models/reset');
 const source = require('../src/models/source');
 const user = require('../src/models/user');
+const utils = require('./utils');
 
-const importDashboards = async () => {
-  const dashboards = JSON.parse(fs.readFileSync(path.join(path.dirname(__filename), 'mock_data/dashboard.json'), {encoding: 'utf-8'}));
-  await dashboard.deleteMany({});
-  await dashboard.insertMany(dashboards);
-};
-const importReset = async () => {
-  const resets = JSON.parse(fs.readFileSync(path.join(path.dirname(__filename), 'mock_data/reset.json'), {encoding: 'utf-8'}));
-  await reset.deleteMany({});
-  await reset.insertMany(resets);
-};
-
-const importSources = async () => {
-  const sources = JSON.parse(fs.readFileSync(path.join(path.dirname(__filename), 'mock_data/source.json'), {encoding: 'utf-8'}));
-  await source.deleteMany({});
-  await source.insertMany(sources);
-};
-
-const importUsers = async () => {
-  const users = JSON.parse(fs.readFileSync(path.join(path.dirname(__filename), 'mock_data/user.json'), {encoding: 'utf-8'}));
-  await user.deleteMany({});
-  await user.insertMany(users);
-};
 test.before(async () => {
   const mongoUrl = process.env.MONGODB_URI;
   await mongoose.connect(mongoUrl, {useNewUrlParser: true, useUnifiedTopology: true});
-  await importUsers();
-  await importDashboards();
-  await importSources();
-  await importReset();
+  await utils.importData();
 });
 
 test.before(async (t) => {
